Unify all related task modules into a single `tasks` module
