+++ tasks/src/commonMain/kotlin/com/taskodoro/tasks/model/Task.kt
@@ -14,18 +14,8 @@
  *    limitations under the License.
  */
 
-package com.taskodoro.tasks.domain
+package com.taskodoro.tasks.model
 
-import com.taskodoro.tasks.model.Task
-import kotlinx.coroutines.flow.Flow
-
-interface GetTasksUseCase {
-    operator fun invoke(): Flow<List<Task>>
-}
-
-class GetTasks(
-    private val loadTasks: () -> Flow<List<Task>>
-) : GetTasksUseCase {
-
-    override fun invoke(): Flow<List<Task>> = loadTasks()
-}
+data class Task(
+    val title: String,
+)
